admin user post and category are done

+++ app/Http/Controllers/AdminUsersController.php
@@ -147,8 +147,9 @@ class AdminUsersController extends Controller
         //
         $user = User::findOrFail($id);
 
-
-        unlink(public_path() . $user->photo->file);
+        if($user->photo_id) {
+            unlink(public_path() . $user->photo->file);
+        }
 
         $user->delete();
